+++ src/vue/modules/movements-history/components/movements-table.vue
@@ -38,7 +38,7 @@
       :movement="movement"
       :key="movement.id"
     />
-    <empty-list-placeholder
+    <empty-tbody-placeholder
       v-if="!movements.length"
     />
   </table>
@@ -46,13 +46,13 @@
 
 <script>
 import MovementsTableRow from './movements-table-row'
-import EmptyListPlaceholder from './movements-empty-list-placeholder.vue'
+import EmptyTbodyPlaceholder from './movements-empty-list-placeholder.vue'
 
 export default {
   name: 'movement-list-renderer',
   components: {
     MovementsTableRow,
-    EmptyListPlaceholder,
+    EmptyTbodyPlaceholder,
   },
   props: {
     movements: {